refactor(transformer): move transformers to utils folder
