+++ packages/sp-proxy/src/interface-adapters/utils/transformers/GenerateMetadataTransformer.ts
